Moved post utility methods from wordprss-update script to wordpress module.
